requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-genai==0.3.0
python-dotenv==1.0.0
